app/requirements.txt
alembic==1.11.1
blinker==1.6.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-MySQLdb==1.0.1
flask-sqlalchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
pycparser==2.21
PyMySQL==1.1.0
pysqlite3==0.5.1
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing-extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
# gunicorn